refactor: modularize authentication layer by moving repository and API definitions into the auth feature package
