+++ app/src/main/java/must/kdroiders/hustlehub/data/repository/UserRepository.kt
@@ -5,9 +5,9 @@ import android.net.Uri
 import dagger.hilt.android.qualifiers.ApplicationContext
 import must.kdroiders.hustlehub.data.model.User
 import must.kdroiders.hustlehub.data.model.UserRole
-import must.kdroiders.hustlehub.data.remote.AuthApiService
+import must.kdroiders.hustlehub.ui.auth.data.remote.AuthApiService
+import must.kdroiders.hustlehub.ui.auth.data.remote.RegisterRequest
 import must.kdroiders.hustlehub.data.remote.MediaApiService
-import must.kdroiders.hustlehub.data.remote.RegisterRequest
 import must.kdroiders.hustlehub.data.remote.UserApiService
 import okhttp3.MediaType.Companion.toMediaTypeOrNull
 import okhttp3.MultipartBody
